Replace OpenAIKit with OpenAIService
